// fronted/src/app/domain/persona/persona_natural/models/persona_natural.model.ts
export class personaNatural {

  idpersonanatural?: number;
  fecha?: string;
  cubano: boolean;
  activo: boolean;
  idpaisorigen: any;
  idpaisresidencia: any;
  correoelectronico: string;
  telefono: string;
  direccion: string;
  primernombre: string;
  segundonombre: string;
  primerapellido: string;
  segundoapellido: string;
  numeroidentidad: string;
  idnraza: any;
  idnsexo: any;
  idnestadocivil: any;
  idprovincia: any;
  idmunicipio: any;
}